better reset game
