+++ src/Contexts/PlayersContext.tsx
@@ -6,6 +6,7 @@ export type PlayersContextType = {
   setPlayers: (players: Player[]) => void;
   updatePlayer: (updatedPlayer: Player) => void;
   updateLifeTotal: (player: Player, updatedLifeTotal: number) => number;
+  resetCurrentGame: () => void;
 };
 
 export const PlayersContext = createContext<PlayersContextType | null>(null);
